Move SumPossibleAcrossRegions into pipeline_utils and setup v2
